\name{EMundirK1Stab-package}
\alias{EMundirK1Stab-package}
\alias{EMundirK1Stab}
\docType{package}
\title{
\packageTitle{EMundirK1Stab}
}
\description{
\packageDescription{EMundirK1Stab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EMundirK1Stab}
\packageIndices{EMundirK1Stab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EMundirK1Stab}

Maintainer: \packageMaintainer{EMundirK1Stab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
